Issue 278: Consolidate caching implementations, MenuModel merged into PageModel.
