+++ README.txt
@@ -2,9 +2,9 @@ README.txt
 
 Welcome to TightBlog, which started off as a fork of Apache Roller in May 2015.  (Prior to starting TightBlog I was a volunteer on the Apache Roller team for 2.5 years.)  Tightblog is not yet ready for release, the Milestone 1.0-tagged Issues must all be complete first.
 
-October 2015 Update:  https://web-gmazza.rhcloud.com/blog/entry/tightblog-taking-shape
+Check my [blog](https://web-gmazza.rhcloud.com/blog/category/Blogs+%26+Wikis) for recent status updates.
 
-As of 26 November 2015, Tightblog uses 18 database tables compared to Roller's 33, and has slimmed down to 293 Java source files from the original 493 in Roller.
+As of 26 November 2015, Tightblog uses 18 database tables compared to Roller's 33, and has slimmed down to 290 Java source files from the 493 in Roller.
 
 Differences between TightBlog and Roller can be seen from the closed issues page:
   https://github.com/gmazza/tightblog/issues?q=is%3Aissue+is%3Aclosed
