+++ app/src/main/java/org/apache/roller/weblogger/util/cache/LazyExpiringCacheEntry.java
@@ -14,8 +14,10 @@
  * limitations under the License.  For additional information regarding
  * copyright in this work, please see the NOTICE file in the top level
  * directory of this distribution.
+ *
+ * Source file modified from the original ASF source; all changes made
+ * are also under Apache License.
  */
-
 package org.apache.roller.weblogger.util.cache;
 
 import java.io.Serializable;
@@ -31,7 +33,7 @@ import java.io.Serializable;
  * entry is "fresh".  If the object is not fresh then we don't return it.
  *
  * This essentially allows us to track when an object is cached and then before
- * we can retrieve that cached object we must compare it with it's last known
+ * we can retrieve that cached object we must compare it with its last known
  * invalidation time to make sure it hasn't expired.  This is useful because
  * instead of actively purging lots of cached objects from the cache at 
  * invalidation time, we can now be lazy and just invalidate them when we
@@ -43,17 +45,14 @@ import java.io.Serializable;
  * we try to use them.
  */
 public class LazyExpiringCacheEntry implements Serializable {
-    
     private Object value = null;
     private long timeCached = -1;
-    
-    
+
     public LazyExpiringCacheEntry(Object item) {
         this.value = item;
         this.timeCached = System.currentTimeMillis();
     }
-    
-    
+
     /**
      * Retrieve the value of this cache entry if it is still "fresh".
      *
@@ -66,19 +65,16 @@ public class LazyExpiringCacheEntry implements Serializable {
             return this.value;
         }
     }
-    
-    
+
     /**
      * Determine if this cache entry has expired.
      */
     public boolean isInvalid(long lastInvalidated) {
-        
         return (this.timeCached < lastInvalidated);
     }
 
-    
     public long getTimeCached() {
         return timeCached;
     }
-    
+
 }
