+++ app/src/main/resources/spring-beans.xml
@@ -122,10 +122,6 @@
     <beans:bean id="messageModel" class="org.apache.roller.weblogger.ui.rendering.model.MessageModel" scope="prototype"/>
     <beans:bean id="calendarModel" class="org.apache.roller.weblogger.ui.rendering.model.CalendarModel" scope="prototype"/>
 
-    <beans:bean id="menuModel" class="org.apache.roller.weblogger.ui.rendering.model.MenuModel" scope="prototype">
-        <beans:property name="userManager" ref="userManager"/>
-    </beans:bean>
-
     <beans:bean id="planetModel" class="org.apache.roller.weblogger.ui.rendering.model.PlanetModel" scope="prototype">
         <beans:property name="urlStrategy" ref="urlStrategy"/>
         <beans:property name="planetManager" ref="planetManager"/>
@@ -166,7 +162,6 @@
                 <beans:ref bean="urlModel"/>
                 <beans:ref bean="messageModel"/>
                 <beans:ref bean="calendarModel"/>
-                <beans:ref bean="menuModel"/>
                 <beans:ref bean="planetModel"/>
             </beans:set>
         </beans:property>
@@ -201,7 +196,6 @@
                 <beans:ref bean="urlModel"/>
                 <beans:ref bean="messageModel"/>
                 <beans:ref bean="calendarModel"/>
-                <beans:ref bean="menuModel"/>
             </beans:set>
         </beans:property>
     </beans:bean>
@@ -216,7 +210,6 @@
                 <beans:ref bean="previewUrlModel"/>
                 <beans:ref bean="messageModel"/>
                 <beans:ref bean="calendarModel"/>
-                <beans:ref bean="menuModel"/>
             </beans:set>
         </beans:property>
     </beans:bean>
