+++ app/src/main/webapp/WEB-INF/velocity/weblog.vm
@@ -47,7 +47,7 @@ These are the documented macros in order defined in this file:
     #showWeblogEntryCalendar($weblog $category)
 
     #showPageMenu($weblog)
-    #showAuthorMenu($vertical)
+    #showEditorMenu($vertical)
 
     #showWeblogSearchForm($weblog)
     #showMobileSearchForm($weblog)
@@ -649,14 +649,14 @@ $calendarModel.showWeblogEntryCalendarBig($weblog, $category)
 
 
 #** 
-Show weblog author menu with vertical or horizontal orientation.
+Show weblog editor menu with vertical or horizontal orientation.
 *#
-#macro(showAuthorMenu $vertical)
+#macro(showEditorMenu $vertical)
   
-  #set( $authorMenu = $menuModel.authorMenu )
-  #if( $authorMenu )
+  #set( $editorMenu = $model.editorMenu )
+  #if( $editorMenu )
       #if( $vertical )
-          #foreach( $menu in $authorMenu.tabs )
+          #foreach( $menu in $editorMenu.tabs )
             <ul class="rMenu">
             #foreach( $item in $menu.items )
                 <li class="rMenuItem"><a href="$url.action($item.action, '/roller-ui/authoring')">$text.get( $menu.key ):$text.get( $item.key )</a></li>
@@ -665,7 +665,7 @@ Show weblog author menu with vertical or horizontal orientation.
           #end
       #else
           #set( $menuCnt = 0 )
-          #foreach( $menu in $authorMenu.tabs )
+          #foreach( $menu in $editorMenu.tabs )
             #foreach( $item in $menu.items )
                 #if ($menuCnt > 0)&nbsp;|&nbsp;#end
                 <a href="$url.action($item.action, '/roller-ui/authoring')">$text.get( $menu.key ):$text.get( $item.key )</a>
@@ -678,37 +678,7 @@ Show weblog author menu with vertical or horizontal orientation.
 #end
 
 
-#** 
-Show weblog admin menu with vertical or horizontal orientation.
-*#
-#macro(showAdminMenu $vertical)
-  
-  #set( $adminMenu = $menuModel.adminMenu )
-  #if( $adminMenu )
-      #if( $vertical )
-          #foreach( $menu in $adminMenu.tabs )
-            <ul class="rMenu">
-            #foreach( $item in $menu.items )
-                <li class="rMenuItem"><a href="$url.action($item.action, '/roller-ui/admin')">$text.get( $menu.key ):$text.get( $item.key )</a></li>
-            #end
-            </ul>
-          #end
-      #else
-          #set( $menuCnt = 0 )
-          #foreach( $menu in $adminMenu.tabs )
-            #foreach( $item in $menu.items )
-                #if ($menuCnt > 0)&nbsp;|&nbsp;#end
-                <a href="$url.action($item.action, '/roller-ui/admin')">$text.get( $menu.key ):$text.get( $item.key )</a>
-                #set($menuCnt = $menuCnt + 1)
-            #end
-          #end
-      #end
-  #end
-
-#end
-
-
-#** 
+#**
 Show search form for a weblog.
 *#
 #macro(showWeblogSearchForm $weblog $showCatChooser)
