+++ app/src/main/webapp/themes/basicmobile/sidebar.vm
@@ -23,5 +23,5 @@
     <h2>Navigation</h2>
     <div class="sidebar">
     #showPageMenu($model.weblog)
-    #showAuthorMenu(true)
+    #showEditorMenu(true)
     </div>
